feat(graph): improve node hover interaction and add GraphVisualizer component
